Fixed issue with validating user in system
Now users are validated as a whole and information is filled in based on what type they are.
Prior to fix, anyone could access veteran and company page. Now site will take them back to home page
.
Also added the models for the company page. Updated the database for copmany.